// packages/backend/src/interfaces/Settings.ts
export interface Settings {
  id?: number;
  headerIcon: string;
  menuIcon: string;
  footerIcon: string;
  cookies_policy: string;
  notifications_policy: string;
  companyName: string;
  companySocialName: string;
  facebookURL: string;
  facebookActive: boolean;
  twitterURL: string;
  twitterActive: boolean;
  instagramURL: string;
  instagramActive: boolean;
  hotmartURL: string;
  hotmartActive: boolean;
  linkedinURL: string;
  linkedinActive: boolean;
}